veux: Disable RescueParty
RescueParty's aggressive approach to error recovery almost never helps
with the problems we're likely to face, and its risk of data loss is far
too high. Force-disable it entirely and remove all other conditions,
especially those that depend on data integrity in the first place, to
reduce the risk of RescueParty triggering falsely and causing data loss.

Note that it is still possible to force-enable RescueParty for testing
purposes using the persist.sys.enable_rescue property.

Change-Id: I72e833e0fcea8833a91448a2eac1a4fd5f65c007
Signed-off-by: tanvirr007 <tanvir.mymnsngh@gmail.com>
Signed-off-by: alternoegraha <noegrahachan@gmail.com>